official-plugins/image-gen/pyproject.toml
[project]
name = "image-gen"
version = "1.0.0"
description = "AI Image Generation Plugin for DawnChat"
requires-python = ">=3.11"
dependencies = [
    "nicegui>=1.4.0",
    "dawnchat-sdk",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
